Refine UPDATE/DELETE alias resolution and fix regression tests
- Improved alias resolution in UPDATE/DELETE to correctly target tables within complex FROM clauses.
- Added support for comma-separated tables (implicit CROSS JOIN) in UPDATE/DELETE FROM.
- Enabled APPLY support in UPDATE/DELETE FROM.
- Refactored DML parser into modular files (insert, update, delete, merge, output) following SRP.
- Fixed multiple regression test failures:
  - AVG precision for INT types.
  - DATEADD formatting and scale.
  - DECLARE parsing in cursor tests.
  - OUTPUT INTO positional column matching.
  - Math unary function result types.
- Added Engine::exec() and Engine::query() helpers for test cleaner code.
- Resolved various compilation errors and unused warnings.
