--- crates/tsql_core/src/executor/value_ops.rs
@@ -1,7 +1,0 @@
-pub mod coercion;
-pub mod comparison;
-pub mod formatting;
-
-pub use coercion::coerce_value_to_type;
-pub use comparison::{compare_values, categorize, truthy, value_key, ValueCategory};
-pub use formatting::{convert_with_style, normalize_datetime_string, format_datetime};
