+++ crates/tsql_core/src/parser/statements/dml/mod.rs
@@ -1,0 +1,10 @@
+pub(crate) mod insert;
+pub(crate) mod update;
+pub(crate) mod delete;
+pub(crate) mod merge;
+pub(crate) mod output;
+
+pub(crate) use insert::*;
+pub(crate) use update::*;
+pub(crate) use delete::*;
+pub(crate) use merge::*;
