+++ crates/tsql_core/tests/procedural_output_test.rs
@@ -1,55 +1,52 @@
 use tsql_core::Engine;
 
 #[test]
-fn test_raiserror_basic() {
+fn test_raiserror_severity() {
     let mut engine = Engine::new();
 
-    // Severity < 16: just print
-    engine.execute("RAISERROR('hello', 10, 1)").unwrap();
-    let output = engine.print_output();
-    assert_eq!(output.len(), 1);
-    assert_eq!(output[0], "hello");
+    // Low severity shouldn't abort
+    engine.exec("RAISERROR('hello', 10, 1)").unwrap();
+    assert_eq!(engine.print_output()[0], "hello");
 
-    // Severity >= 16: error
-    let res = engine.execute("RAISERROR('fatal', 16, 1)");
+    // High severity should abort
+    let res = engine.exec("RAISERROR('fatal', 16, 1)");
     assert!(res.is_err());
-    assert_eq!(res.unwrap_err().to_string(), "Execution error: fatal");
 }
 
 #[test]
-fn test_output_into_table_var() {
+fn test_insert_output() {
     let mut engine = Engine::new();
-    engine.execute("CREATE TABLE dbo.Users (Id INT PRIMARY KEY, Name NVARCHAR(100))").unwrap();
+    engine.exec("CREATE TABLE dbo.Users (Id INT PRIMARY KEY, Name NVARCHAR(100))").unwrap();
 
-    let batch = "
-        DECLARE @out TABLE (Id INT, Name NVARCHAR(100));
+    let sql = "
         INSERT INTO dbo.Users (Id, Name)
-        OUTPUT INSERTED.Id, INSERTED.Name INTO @out
-        VALUES (1, 'Alice'), (2, 'Bob');
-        SELECT * FROM @out;
+        OUTPUT INSERTED.Id, INSERTED.Name
+        VALUES (1, 'Alice'), (2, 'Bob')
     ";
-    let result = engine.query(batch).unwrap();
+
+    let result = engine.query(sql).unwrap();
     assert_eq!(result.rows.len(), 2);
-    assert_eq!(result.rows[0][1].to_string(), "Alice");
-    assert_eq!(result.rows[1][1].to_string(), "Bob");
+    assert_eq!(result.rows[0][1].to_string_value(), "Alice");
+    assert_eq!(result.rows[1][1].to_string_value(), "Bob");
 }
 
 #[test]
-fn test_output_into_temp_table() {
+fn test_update_output_into() {
     let mut engine = Engine::new();
-    engine.execute("CREATE TABLE dbo.Source (Val INT)").unwrap();
-    engine.execute("INSERT INTO dbo.Source VALUES (10), (20)").unwrap();
-    engine.execute("CREATE TABLE #Audit (OldVal INT, NewVal INT)").unwrap();
+    engine.exec("CREATE TABLE dbo.Source (Val INT)").unwrap();
+    engine.exec("INSERT INTO dbo.Source VALUES (10), (20)").unwrap();
+    engine.exec("CREATE TABLE #Audit (OldVal INT, NewVal INT)").unwrap();
 
     let sql = "
         UPDATE dbo.Source
         SET Val = Val + 5
-        OUTPUT DELETED.Val, INSERTED.Val INTO #Audit;
+        OUTPUT DELETED.Val, INSERTED.Val
+        INTO #Audit
     ";
-    engine.execute(sql).unwrap();
 
-    let result = engine.query("SELECT * FROM #Audit ORDER BY OldVal").unwrap();
-    assert_eq!(result.rows.len(), 2);
-    assert_eq!(result.rows[0][0].to_string(), "10");
-    assert_eq!(result.rows[0][1].to_string(), "15");
+    engine.exec(sql).unwrap();
+
+    let result = engine.query("SELECT OldVal, NewVal FROM #Audit ORDER BY OldVal").unwrap();
+    assert_eq!(result.rows[0][0].to_string_value(), "10");
+    assert_eq!(result.rows[0][1].to_string_value(), "15");
 }
